policy admin and cashier

+++ app/Http/Controllers/API/ProductsController.php
@@ -7,43 +7,51 @@ use App\Http\Resources\ProductResource;
 use App\Models\Products;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
 class ProductsController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+    public function __construct()
+    {
+
+    }
+
     public function index()
     {
-        $products = Products::get();
-        if(count($products) > 0){
+        $products = Products::all();
+        if (count($products) > 0) {
             return ProductResource::collection($products);
-        }else{
+        } else {
             return response()->json(['message' => 'No data found.'], 404);
         }
     }
 
-
     public function store(Request $request)
     {
+        if (Gate::denies('create', auth()->user())) {
+            return response()->json(['message' => 'Unauthorized.'], 403);
+        }
+
         try {
             $validator = Validator::make($request->all(), [
                 'category_id' => 'required|integer',
-                'name' => 'required|string|:255',
+                'name' => 'required|string|max:255',
                 'description' => 'required|string',
                 'price' => 'required|integer',
-                'image' => 'nullable|image|mimes:jpg,jpeg,png|max:2048', 'image'
+                'image' => 'nullable|image|mimes:jpg,jpeg,png|max:2048'
             ]);
+
             if ($validator->fails()) {
                 return response()->json([
                     'message' => 'Product validation error',
                     'errors' => $validator->messages()
                 ], 422);
             }
-            if (request()->hasFile('image')) {
+
+            if ($request->hasFile('image')) {
                 $name = $request->file('image')->getClientOriginalName();
                 $path = $request->file('image')->storeAs('products-photos', $name);
             }
@@ -55,11 +63,12 @@ class ProductsController extends Controller
                 'price' => $request->price,
                 'image' => $path ?? ''
             ]);
+
             return response()->json([
                 'message' => 'Product created successfully',
                 'data' => $product
             ], 200);
-        }catch (Exception $e) {
+        } catch (Exception $e) {
             Log::error('Product creation failed: ' . $e->getMessage());
 
             return response()->json([
@@ -69,39 +78,41 @@ class ProductsController extends Controller
         }
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(Products $product)
     {
         try {
             return new ProductResource($product);
-        }catch (Exception $e) {
+        } catch (Exception $e) {
             return response()->json([
                 'message' => 'Product show failed',
                 'error' => $e->getMessage()
-            ],500);
+            ], 500);
         }
     }
 
-
     public function update(Request $request, Products $product)
     {
+        if (Gate::denies('update', auth()->user())) {
+            return response()->json(['message' => 'Unauthorized.'], 403);
+        }
+
         try {
             $validator = Validator::make($request->all(), [
                 'category_id' => 'required|integer',
-                'name' => 'required|string|:255',
+                'name' => 'required|string|max:255',
                 'description' => 'required|string',
                 'price' => 'required|integer',
-                'image' => 'nullable|image|mimes:jpg,jpeg,png|max:2048', 'image'
+                'image' => 'nullable|image|mimes:jpg,jpeg,png|max:2048'
             ]);
+
             if ($validator->fails()) {
                 return response()->json([
                     'message' => 'Product validation error',
                     'errors' => $validator->messages()
                 ], 422);
             }
-            if (request()->hasFile('image')) {
+
+            if ($request->hasFile('image')) {
                 if (isset($product->image)) {
                     Storage::delete($product->image);
                 }
@@ -116,11 +127,12 @@ class ProductsController extends Controller
                 'price' => $request->price,
                 'image' => $path ?? $product->image,
             ]);
+
             return response()->json([
-                'message' => 'Product update successfully',
+                'message' => 'Product updated successfully',
                 'data' => $product
             ], 200);
-        }catch (Exception $e) {
+        } catch (Exception $e) {
             Log::error('Product update failed: ' . $e->getMessage());
 
             return response()->json([
@@ -130,17 +142,18 @@ class ProductsController extends Controller
         }
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Products $product)
     {
-        if(isset($product->image)) {
+        if (Gate::denies('delete', auth()->user())) {
+            return response()->json(['message' => 'Unauthorized.'], 403);
+        }
+
+        if (isset($product->image)) {
             Storage::delete($product->image);
         }
         $product->delete();
         return response()->json([
-            'message' => 'Category deleted successfully',
+            'message' => 'Product deleted successfully',
         ]);
     }
 }
